Added controls for errors in php side

+++ HUNGRY.bo/PHP/supplier_home.php
@@ -54,7 +54,7 @@ if(isset($_GET["request"])) {
           $stmt->close();
 
         }
-        
+
         if(isset($_POST["immagine"])) {
           $immagine = $_POST["immagine"];
           $stmt = $mysqli->prepare("UPDATE Fornitore SET Immagine=? WHERE Username=?");
@@ -79,27 +79,31 @@ if(isset($_GET["request"])) {
         break;
 
       case 'modifica-orari':
-        $apertura = $_POST["open"];
-        $chiusura = $_POST["close"];
+        if(isset($_POST["open"]) && isset($_POST["close"])) {
+          $apertura = $_POST["open"];
+          $chiusura = $_POST["close"];
 
-        $stmt = $mysqli->prepare("UPDATE Fornitore SET OraApertura=?, OraChiusura=? WHERE Username=?");
+          $stmt = $mysqli->prepare("UPDATE Fornitore SET OraApertura=?, OraChiusura=? WHERE Username=?");
 
-        if($stmt == false) {
-          $response_array['status'] = "Errore nella query";
+          if($stmt == false) {
+            $response_array['status'] = "Errore nella query";
+            print json_encode($response_array);
+            die();
+          }
+
+          $stmt->bind_param('sss', $apertura, $chiusura, $_SESSION['username']);
+
+          $stmt->execute();
+
+          $stmt->close();
+
+          $response_array['status'] = 'success';
+
           print json_encode($response_array);
-          die();
+        } else {
+          $response_array['status'] = "Orari non inseriti";
+          print json_encode($response_array);
         }
-
-        $stmt->bind_param('sss', $apertura, $chiusura, $_SESSION['username']);
-
-        $stmt->execute();
-
-        $stmt->close();
-
-        $response_array['status'] = 'success';
-
-        print json_encode($response_array);
-
         break;
 
       case 'aggiungi-prodotto':
@@ -179,24 +183,28 @@ if(isset($_GET["request"])) {
         break;
 
       case 'rimuovi-prodotto':
-        $stmt = $mysqli->prepare("DELETE FROM Prodotto WHERE ID=?");
+        if(isset($_POST['id'])) {
+          $stmt = $mysqli->prepare("DELETE FROM Prodotto WHERE ID=?");
 
-        if($stmt == false) {
-          $response_array['status'] = "Errore nella query";
+          if($stmt == false) {
+            $response_array['status'] = "Errore nella query";
+            print json_encode($response_array);
+            die();
+          }
+
+          $stmt->bind_param('i', $_POST['id']);
+
+          $stmt->execute();
+
+          $stmt->close();
+
+          $response_array['status'] = 'success';
+
           print json_encode($response_array);
-          die();
+        } else {
+          $response_array['status'] = "ID mancante";
+          print json_encode($response_array);
         }
-
-        $stmt->bind_param('i', $_POST['id']);
-
-        $stmt->execute();
-
-        $stmt->close();
-
-        $response_array['status'] = 'success';
-
-        print json_encode($response_array);
-
         break;
 
       case 'controllo-notifiche':
@@ -252,72 +260,87 @@ if(isset($_GET["request"])) {
         break;
 
       case 'conta-prodotti':
-        $stmt = $mysqli->prepare("SELECT COUNT(*) FROM ProdottoInOrdine WHERE IDOrdine = ?");
+        if(isset($_POST['id'])) {
+          $stmt = $mysqli->prepare("SELECT COUNT(*) FROM ProdottoInOrdine WHERE IDOrdine = ?");
 
-        if($stmt == false) {
-          $response_array['status'] = "Errore nella query";
+          if($stmt == false) {
+            $response_array['status'] = "Errore nella query";
+            print json_encode($response_array);
+            die();
+          }
+
+          $stmt->bind_param('i', $_POST['id']);
+
+          $stmt->execute();
+
+          $stmt->bind_result($result);
+
+          $stmt->fetch();
+
+          if($result > 0) {
+            $response_array['count'] = $result;
+          } else {
+            $response_array['status'] = 'false';
+          }
+
           print json_encode($response_array);
-          die();
-        }
-
-        $stmt->bind_param('i', $_POST['id']);
-
-        $stmt->execute();
-
-        $stmt->bind_result($result);
-
-        $stmt->fetch();
-
-        if($result > 0) {
-          $response_array['count'] = $result;
         } else {
-          $response_array['status'] = 'false';
+          $response_array['status'] = "ID mancante";
+          print json_encode($response_array);
         }
-
-        print json_encode($response_array);
-
         break;
 
       case 'ordine-notifica':
-        $stmt = $mysqli->prepare("SELECT * FROM Ordine WHERE ID = ?");
+        if(isset($_POST['id'])) {
+          $stmt = $mysqli->prepare("SELECT * FROM Ordine WHERE ID = ?");
 
-        if($stmt == false) {
-          $response_array['status'] = "Errore nella query";
+          if($stmt == false) {
+            $response_array['status'] = "Errore nella query";
+            print json_encode($response_array);
+            die();
+          }
+
+          $stmt->bind_param('i', $_POST['id']);
+
+          $stmt->execute();
+
+          $result = $stmt->get_result();
+
+          $output = array();
+          while($row = $result->fetch_assoc()){
+              $output[] = $row;
+          }
+          $stmt->close();
+
+          print json_encode($output);
+        } else {
+          $response_array['status'] = "ID mancante";
           print json_encode($response_array);
-          die();
         }
-
-        $stmt->bind_param('i', $_POST['id']);
-
-        $stmt->execute();
-
-        $result = $stmt->get_result();
-
-        $output = array();
-        while($row = $result->fetch_assoc()){
-            $output[] = $row;
-        }
-        $stmt->close();
-
-        print json_encode($output);
         break;
 
       case 'rimuovi-notifica':
-        $stmt = $mysqli->prepare("DELETE FROM Notifica WHERE ID=?");
+        
+        if(isset($_POST['id'])) {
+          $stmt = $mysqli->prepare("DELETE FROM Notifica WHERE ID=?");
 
-        if($stmt == false) {
-          $response_array['status'] = "Errore nella query";
+          if($stmt == false) {
+            $response_array['status'] = "Errore nella query";
+            print json_encode($response_array);
+            die();
+          }
+
+          $stmt->bind_param('i', $_POST['id']);
+
+          $stmt->execute();
+
+          $response_array['status'] = "success";
+
           print json_encode($response_array);
-          die();
+        } else {
+          $response_array['status'] = "ID mancante";
+          print json_encode($response_array);
         }
-
-        $stmt->bind_param('i', $_POST['id']);
-
-        $stmt->execute();
-
-        $response_array['status'] = "success";
-
-        print json_encode($response_array);
         break;
     }
     $mysqli->close();
